feat(inspector): Add ComponentRef.cs class to store references to scene components Add new attribute TransformFormatterAttribute.cs to store references to scene transforms Changed deltaTime type from double to float Add TimeManager.cs for global access to deltaTime and Application.cs play time

+++ FlyEngine.Core/Engine/Behaviours/CameraController.cs
@@ -29,7 +29,7 @@ public class CameraController : Behaviour
         Input.CursorVisible = false;
     }
 
-    public override void OnUpdate(double deltaTime)
+    public override void OnUpdate(float deltaTime)
     {
         _rotation.X += Input.MouseInput.Y * Sensitivity;
         _rotation.Y -= Input.MouseInput.X * Sensitivity;
